refactor: Enable Telemetry for CLI
