+++ extension/src/Settings/SettingsLoader.ts
@@ -1,6 +1,4 @@
 import * as vscode from "vscode";
-import * as path from "path";
-import * as fs from "fs";
 import {
   AppManifest,
   IAppSourceCopSettings,
@@ -85,14 +83,5 @@ function getWorkspaceFolderPath(): string {
 }
 
 export function getExtensionPackage(): IExtensionPackage {
-  let filePath = path.resolve(__dirname, "..", "package.json");
-  if (!fs.existsSync(filePath)) {
-    // Debugging, with another file structure because of not using webpack
-    filePath = path.resolve(__dirname, "..", "..", "package.json");
-  }
-
-  const extensionPackage = JSON.parse(
-    fs.readFileSync(filePath, "utf8")
-  ) as IExtensionPackage;
-  return extensionPackage;
+  return CliSettingsLoader.getExtensionPackage();
 }
